deleted table tasker and header footer tasker
