+++ labeling/train_dataset/api/web/form_input_archive.html
@@ -68,9 +68,7 @@
                                 <option value="law_classifier" selected>law</option>
                                 <option value="tz_classifier">tz</option>
                                 <option value="diploma_classifier">diploma</option>
-                                <option value="header_classifier">header_classifier</option>
                                 <option value="paragraph_classifier">paragraph</option>
-                                <option value="tables_classifier">tables</option>
                             </select> type_of_task
                         </label>
                     </p>
